# swagger/swagger.yaml
...
openapi: 3.0.1
info:
  title: Smart CRM
  description: API Documentation for Smart CRM Project by JDUAR NVIDIA Team
  termsOfService: 'http://swagger.io/terms/'
  version: 1.0.0
servers:
  - url: 'https://ppl-smartcrm-backend.herokuapp.com/'
...
paths:
...
  '/crossroads/registpasscode':
    post:
      security:
      - OAuth2: []
      tags:
        - Registration
      summary: Send registration photo to API Smart CRM XQ
      operationId: register_passcode
      description: Passcode chosen from frontend is sent to backend in a string id. Backend validate the paylod and send it to API Smart CRM XQ.
      requestBody:
        description: Passcode chosen by user.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChosenPasscode'

        required: true
      responses:
        '200':
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterSuccess'
        '400':
          description: Invalid Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvalidRequest'
        '403':
          description: Forbidden Access
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Forbidden'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
...
components:
  securitySchemes:
    OAuth2:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: https://example.com/oauth/authorize
          tokenUrl: https://example.com/oauth/token
          scopes:
            admin: Grants access to admin operations
  schemas:
...
    ChosenPasscode:
      type: object
      properties:
        chosen_passcode:
          type: string
          enum:
            - "Grapes"
            - "Melon"
            - "Watermelon"
            - "Tangerine"
            - "Lemon"
            - "Banana"
            - "Pineapple"
            - "Mango"
            - "Apple"
            - "Pear"
            - "Peach"
            - "Cherries"
            - "Strawberry"
            - "Kiwi"
            - "Tomato"
            - "Coconut"
            - "Avocado"
            - "Eggplant"
            - "Cucumber"
          example: "Grapes"
    RegisterSuccess:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: "User registration successful"
...
    Forbidden:
      type: object
      properties:
        code:
          type: integer
          example: 403
        message:
          type: string
          example: Forbidden Access
    InvalidRequest:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          description: Terjadi ketika email/password tidak sesuai format penulisan. Invalid Request terjadi ketika payload tidak sesuai format.
          example: "Invalid Request"
    ApiErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 500
        message:
          type: string
          enum:
          - "Internal Server: Create Payload Failed"
          - "Internal Server: Invalid Payload"
          example: "Internal Server: Invalid Payload"